Rename tests to just account for Notification surface
